change float to 32

+++ lib/network.py
@@ -86,7 +86,8 @@ class R2N2:
         self.softmax = tf.nn.softmax(self.logits)
         self.log_softmax = tf.log(self.softmax)
         self.cross_entropy = - \
-            tf.multiply(tf.cast(self.label, tf.float64), self.log_softmax)
+            tf.multiply(tf.cast(self.label, self.log_softmax.dtype),
+                        self.log_softmax)
 
         self.losses = tf.reduce_mean(self.cross_entropy, axis=[1, 2, 3])
         self.batch_loss = tf.reduce_mean(self.losses)
